Add a layout for mobiles.
This modification was not requested but helps to test the functionality in the app. It's easier to work with mobiles than with the boxes.
NOTE: this commit might a build error because there is a reference in MainActivity to a component that doesn't exist `fileList = findViewById(R.id.fileList);` in the new UI. But this is fixed in the next commit, which is more important.