Exclude file from auto-connect build
I forgot to add build.gradle which includes a rule to exclude the file `DnsSdManager.java`. This file is going to be deleted once the feature is complete. It was the first attempt to implement DNS-SD. It's not complete and  turned out that the steps I followed are for a different situation (when devices are already in a LAN), they don't apply to P2P.